chore: rename show --> hide
