+++ src/components/universal/core/Hide/Hide.stories.tsx
@@ -1,32 +1,32 @@
-import { Image, Show } from "components/universal";
+import { Image, Hide } from "components/universal";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-type Story = StoryObj<typeof Show>;
+type Story = StoryObj<typeof Hide>;
 
-export const Above: Story = {
+export const HideBelow: Story = {
   render: () => (
     // TODO: figure out why hideBelow seems to respond to the wrong breakpoint
-    <Show hideBelow="sm">
+    <Hide hideBelow="sm">
       <Image src="/img/logo.png" alt="AR logo" height={40} width={40} />
-    </Show>
+    </Hide>
   ),
 };
 
-export const Below: Story = {
+export const HideFrom: Story = {
   render: () => (
-    <Show hideFrom="md">
+    <Hide hideFrom="md">
       <Image src="/img/logo.png" alt="AR logo" height={40} width={40} />
-    </Show>
+    </Hide>
   ),
 };
 
 // TODO remove explicit type annotation, required due to `pnpm` bug (and therefore Yarn with `pnpm` linker);
-const meta: Meta<typeof Show> = {
-  title: "Components/Universal/Core/Show",
-  component: Show,
+const meta: Meta<typeof Hide> = {
+  title: "Components/Universal/Core/Hide",
+  component: Hide,
   tags: ["autodocs"],
   decorators: [(Story) => <Story />],
-} as Meta<typeof Show>;
+} as Meta<typeof Hide>;
 
 export default meta;
