+++ src/components/universal/core/index.ts
@@ -1,7 +1,7 @@
 export { default as Badge } from "./Badge/Badge";
 export { default as Card, CardHeader, CardBody, CardFooter } from "./Card/Card";
+export { default as Hide } from "./Hide/Hide";
 export { default as Image } from "./Image/Image";
-export { default as Show } from "./Show/Show";
 export { default as Spinner } from "./Spinner/Spinner";
 export { default as Text } from "./Text/Text";
 
@@ -11,7 +11,7 @@ export type {
   CardBodyProps,
   CardFooterProps,
 } from "./Card/Card";
+export type { Props as HideProps } from "./Hide/Hide";
 export type { Props as ImageProps } from "./Image/Image";
-export type { Props as ShowProps } from "./Show/Show";
 export type { Props as SpinnerProps } from "./Spinner/Spinner";
 export type { Props as TextProps } from "./Text/Text";
